:recycle: refactor : rename files of constant values to end with Const
