+++ src/main/java/com/run_us/server/domains/running/service/util/RunningKeyUtil.java
@@ -1,13 +1,13 @@
 package com.run_us.server.domains.running.service.util;
 
-import com.run_us.server.domains.running.domain.RunningConstants;
+import com.run_us.server.domains.running.domain.RunningConst;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class RunningKeyUtil {
   public static String createLiveKey(String runningId, String userId, String suffix) {
-    return RunningConstants.RUNNING_PREFIX + runningId + ":" + userId + suffix;
+    return RunningConst.RUNNING_PREFIX + runningId + ":" + userId + suffix;
   }
 
   public static String extractUserIdFromKey(String key) {
