+++ src/main/java/com/run_us/server/global/common/GlobalConst.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class GlobalConsts {
+public final class GlobalConst {
   public static final String TIME_ZONE_ID = "Asia/Seoul";
   public static final String DEFAULT_IMG_URL = "default_img_url";
