+++ src/main/java/com/run_us/server/global/config/WebSocketConfig.java
@@ -10,7 +10,7 @@ import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBr
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 
-import static com.run_us.server.global.common.GlobalConsts.*;
+import static com.run_us.server.global.common.GlobalConst.*;
 
 /**
  * websocket config 클래스
